requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
pydantic==2.8.2
matplotlib
matplotlib==3.9.2
websockets==13.0.1
bcrypt==4.1.2
PyJWT==2.8.0
